feat(web): serve Job detail at /jobs/{jobId}
Canonical Job URL so product links and emails no longer depend on
nested Agent routes. Old /agents/{id}/jobs/{jobId} paths redirect.

Refs SOK-806
