+++ apps/web/src/app/(app)/jobs/[jobId]/loading.tsx
@@ -1,0 +1,5 @@
+import DefaultLoading from "@/components/default-loading";
+
+export default function JobDetailsLoading() {
+  return <DefaultLoading className="h-full min-h-[300px] w-full flex-1 p-8" />;
+}
