+++ apps/web/src/hooks/use-job-submission.ts
@@ -15,6 +15,7 @@ import {
   prepareInputValues,
 } from "@/lib/job-input";
 import type { CoreAgentDto } from "@/lib/types/core-dto";
+import { buildJobHref } from "@/lib/utils/job-href";
 import {
   getUserFileUploadErrorMessage,
   uploadInputDataFiles,
@@ -88,7 +89,7 @@ export function useJobSubmission({
             jobId: result.value.jobId,
           });
           onSuccess();
-          router.push(`/agents/${agentId}/jobs/${result.value.jobId}`);
+          router.push(buildJobHref(result.value.jobId));
         } else {
           switch (result.error.code) {
             case CommonErrorCode.UNAUTHENTICATED:
