+++ apps/web/src/lib/actions/job/__tests__/action.test.ts
@@ -151,8 +151,9 @@ describe("startJob", () => {
       "user-1",
       "ada@example.com",
     );
+    expect(revalidatePath).toHaveBeenCalledWith("/jobs/job-1", "layout");
     expect(revalidatePath).toHaveBeenCalledWith(
-      "/agents/agent-1/jobs/job-1",
+      "/agents/agent-1/jobs",
       "layout",
     );
     expect(result).toEqual({
