+++ apps/web/src/lib/actions/job/action.ts
@@ -26,6 +26,7 @@ import {
   startJobInputSchema,
 } from "@/lib/schemas";
 import { callAgentHiredWebHook, jobService } from "@/lib/services";
+import { buildJobHref } from "@/lib/utils/job-href";
 import { normalizeOptionalProjectId } from "@/lib/utils/project";
 import {
   type AuthenticatedRequest,
@@ -326,7 +327,8 @@ export const startJob = withSession<
 
       // call after agent hired webhook
       void callAgentHiredWebHook(userId, session.user.email);
-      revalidatePath(`/agents/${input.agentId}/jobs/${job.data.id}`, "layout");
+      revalidatePath(buildJobHref(job.data.id), "layout");
+      revalidatePath(`/agents/${input.agentId}/jobs`, "layout");
       return toActionResult(ok({ jobId: job.data.id }));
     } catch (error) {
       scope.setTag("error_type", "job_start_error");
@@ -562,7 +564,7 @@ export const moveJobToWorkspace = withSession<
   try {
     await jobService.moveJobToWorkspace(jobId, organizationId);
     revalidatePath(`/agents/${agentId}/jobs`);
-    revalidatePath(`/agents/${agentId}/jobs/${jobId}`);
+    revalidatePath(buildJobHref(jobId));
     return { jobId };
   } catch (error) {
     console.error("Failed to move job to workspace", error);
